+++ src/ts/AdUnits/VastAdUnit.ts
@@ -40,7 +40,6 @@ export class VastAdUnit extends VideoAdUnit<VastCampaign> {
     private _volume: number;
     private _muted: boolean = false;
     private _events: Array<[number, string]> = [[0.0, 'AdVideoStart'], [0.25, 'AdVideoFirstQuartile'], [0.5, 'AdVideoMidpoint'], [0.75, 'AdVideoThirdQuartile']];
-    private _realDuration: number;
     private _vastCampaign: VastCampaign;
     private _vastPlacement: Placement;
 
@@ -111,14 +110,6 @@ export class VastAdUnit extends VideoAdUnit<VastCampaign> {
         this._events = events;
     }
 
-    public getRealDuration() {
-        return this._realDuration;
-    }
-
-    public setRealDuration(duration: number) {
-        this._realDuration = duration;
-    }
-
     public getVolume() {
         if(this._muted) {
             return 0;
@@ -136,10 +127,6 @@ export class VastAdUnit extends VideoAdUnit<VastCampaign> {
 
     public getMuted() {
         return this._muted;
-    }
-
-    public getDuration(): number | null {
-        return this._vastCampaign.getVast().getDuration();
     }
 
     public sendImpressionEvent(sessionId: string, sdkVersion: number): void {
@@ -247,8 +234,8 @@ export class VastAdUnit extends VideoAdUnit<VastCampaign> {
 
     private sendQuartileEvent(sessionId: string, sdkVersion: number, position: number, oldPosition: number, quartile: number, quartileEventName: string) {
         if (this.getTrackingEventUrls(quartileEventName)) {
-            const duration = this.getDuration();
-            if (duration && duration > 0 && position / 1000 > duration * 0.25 * quartile && oldPosition / 1000 < duration * 0.25 * quartile) {
+            const duration = this._vastCampaign.getVideo().getDuration();
+            if (duration && duration > 0 && position > duration * 0.25 * quartile && oldPosition < duration * 0.25 * quartile) {
                 this.sendTrackingEvent(quartileEventName, sessionId, sdkVersion);
             }
         }