+++ src/ts/Test/Unit/EventHandlers/MRAIDEventHandlersTest.ts
@@ -136,7 +136,7 @@ describe('MRAIDEventHandlersTest', () => {
                     headers: [['location', 'market://foobar.com']]
                 }));
 
-                mraidView = new MRAID(nativeBridge, placement, mraidCampaign);
+                mraidView = new MRAID(nativeBridge, placement, mraidCampaign, false);
                 sinon.stub(mraidView, 'createMRAID').callsFake(() => {
                     return Promise.resolve();
                 });